Fix: Issue with Context Menu.
If you're viewing App in the mobile view, you might notice that if you click into a field which has text and then custom directiveSelectAll is fired, a context menu is appeared for mobile devices.
And this is because this is something that happens on mobile devices. But this can be a little bit annoying when you're debugging your app on a desktop.